+++ src/ui-spa/src/components/case-registration/case-registration-confirmation/index.tsx
